Rename mantle's encoding to codec, to hold common types.
